refactor: safary telemetry
