+++ src/providers/telemetry/capturePostActionEvent.ts
@@ -2,8 +2,8 @@ import { type SocialSource, Source } from '@/constants/enum.js';
 import { UnreachableError } from '@/constants/error.js';
 import { createLookupTableResolver } from '@/helpers/createLookupTableResolver.js';
 import { runInSafe } from '@/helpers/runInSafe.js';
-import { getPostEventParameters } from '@/providers/safary/getPostEventParameters.js';
-import { SafaryTelemetryProvider } from '@/providers/safary/Telemetry.js';
+import { getPostEventParameters } from '@/providers/telemetry/getPostEventParameters.js';
+import { SafaryTelemetryProvider } from '@/providers/telemetry/Safary.js';
 import type { Post } from '@/providers/types/SocialMedia.js';
 import { EventId } from '@/providers/types/Telemetry.js';
