+++ src/providers/telemetry/getPublicParameters.ts
